requirements.txt
absl-py==2.2.0
accelerate==1.2.1
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
astunparse==1.6.3
async-timeout==5.0.1
attrs==24.3.0
beautifulsoup4==4.12.3
bitsandbytes==0.45.0
blis==0.7.11
boltons==21.0.0
cachetools==5.5.0
catalogue==2.0.10
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.21.0
cltk==1.4.0
confection==0.1.5
ConfigArgParse==1.7
configobj==5.0.9
cryptography==45.0.6
cymem==2.0.11
datasets==3.2.0
dill==0.3.8
distro==1.9.0
emoji==2.14.1
et_xmlfile==2.0.0
evaluate==0.4.3
filelock==3.16.1
flatbuffers==25.2.10
frozenlist==1.5.0
fsspec==2024.9.0
gast==0.6.0
gensim==4.3.3
gitdb==4.0.12
GitPython==3.1.44
globus-sdk==3.62.0
google-pasta==0.2.0
greek-accentuation==1.2.0
grpcio==1.71.0
h5py==3.13.0
huggingface-hub==0.27.1
idna==3.10
importlib_metadata==8.6.1
Jinja2==3.1.5
jiwer==3.1.0
joblib==1.4.2
langcodes==3.5.0
language_data==1.3.0
Levenshtein==0.27.1
libclang==18.1.1
lightning-utilities==0.11.9
lxml==5.3.0
marisa-trie==1.2.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mkl-service==2.4.0
ml_dtypes==0.5.1
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
murmurhash==1.0.12
mwparserfromhell==0.6.6
namex==0.0.8
networkx==3.2.1
nltk==3.9.1
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-ml-py==12.535.161
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
nvitop==1.3.2
openpyxl==3.1.5
opt_einsum==3.4.0
optree==0.14.1
packaging==24.2
pandas==2.2.3
parsedatetime==2.6
peft==0.14.0
Pillow==9.0.1
preshed==3.0.9
propcache==0.2.1
protobuf==3.20.3
psutil==6.1.1
pyarrow==18.1.0
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
Pygments==2.19.1
PyJWT==2.10.1
pyRFC3339==2.0.1
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
RapidFuzz==3.12.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
sacremoses==0.1.1
safetensors==0.5.0
scikit-learn==1.5.2
scipy==1.12.0
sentencepiece==0.1.96
six==1.17.0
smart-open==7.1.0
smmap==5.0.2
soupsieve==2.6
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
stanza==1.9.2
stringcase==1.2.0
sympy==1.13.1
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.5.0
tf_keras==2.19.0
thinc==8.2.5
threadpoolctl==3.5.0
tokenizers==0.21.0
tomli==2.2.1
torch==2.5.1
torchaudio
torchmetrics==1.6.1
torchvision==0.20.1
tqdm==4.67.1
transformers==4.47.1
triton==3.1.0
typer==0.4.1
typing-inspection==0.4.0
tzdata==2024.2
Unidecode==1.3.8
urllib3==2.3.0
wasabi==1.1.3
weasel==0.4.1
Werkzeug==3.1.3
wrapt==1.17.2
xxhash==3.5.0
yarl==1.18.3
zipp==3.21.0
